+++ examples/main.js
@@ -1,0 +1,3 @@
+console.log('Hi!');
+
+console.log('d3Kit', d3Kit.helper);