+++ src/main.js
@@ -1,0 +1,6 @@
+
+export { default as Skeleton } from './skeleton.js';
+export { default as Chartlet } from './chartlet.js';
+export { default as factory } from './factory.js';
+export { default as LayerOrganizer } from './layerOrganizer.js';
+export { default as helper } from './helper.js';