+++ CHANGELOG.md
@@ -9,7 +9,7 @@ Make d3Kit compatible with D3 v4. Key changes are due to:
 - Removal of `d3.functor` and `d3.rebind`. Implement helper functions as replacements.
 - API changes for `d3.dispatch`. Now use `dispatch.call('x', ...)` instead of `dispatch.x(...)`
 
-The npm package also moves D3 to `peerDependencies` instead of `dependencies`.
+The npm package also remove `d3` from `dependencies` and add `d3-selection` and `d3-dispatch` to `peerDependencies` instead.
 
 ## v1.x.x